Renamed the temp file to cluster_service_2_x_behaviors.py
